fix(codacy): split test_gui_controller_coverage.py to clear Lizard medium-NLOC
Hoist the shared harness/view/row fixtures into tests/_gui_controller_fixtures.py
and split the controller coverage suite into two focused files:
- test_gui_controller_coverage.py: basic state/selection branches.
- test_gui_controller_operations_coverage.py: operations + boot-state branches.

Both files re-import from the fixtures module instead of duplicating ~250
lines of helpers, keeping each test module under Lizard's medium NLOC
budget so the file-size finding stops firing.
